Fix unification between interface and implementations
When an Bindable<T> is created where T is an subclass of W, you could not
use it in axioms that use Bindable<W>. That is because generics do not inherit
super class like arrays. This commit changes ApplySubstitution method to create
new Bindable<W> when Bindable<T> is passed to the new axiom if T implements W.

Also moved Test files to Example folder since they were tightly coupled with Test data.